stk: Fixed all known issue
Fixed where video sticker cannot be played due to wrong canvas size.
The issue is come from GetImageSize() where the actual image size of
animated webp is located at ANMF. But this is not the best solution
yet because it just one frame, not the entire canvas.

Fixed where video sticker is not playing infinitely due to loop_count
in the ANIM chunk is set to 1 by default by ffmpeg. Setting it to 0
fixes it.

Fixed where sticker metadata is not showing. The issue is come from me,
yes, myself, has a wrong concept about byte ordering. So, the entry tag
should be "WA" but due to wrong byte ordering, it has to be "AW" instead
and cannot be read by the app. Reviewing the code and apply some patches
related to byte ordering issue fixes it.

Signed-off-by: Aziz Ridhwan Pratama <azizpratama2196@gmail.com>